requirements.txt
Flask==2.0.1
python-binance==1.0.12
requests==2.25.1
python-dotenv==0.18.0

